started to add a Tool and Component class
