+++ ctapipe/core/tool.py
@@ -1,0 +1,4 @@
+from traitlets.config import Application
+
+class Tool(Application):
+    pass
